Numeric.Half.Internal

Description

Half-precision floating-point values. These arise commonly in GPU work and it is useful to be able to compute them and compute with them on the CPU as well.

Documentation

newtype Half

A half-precision floating point value

Constructors

Half

Fields

getHalf :: CUShort

isZero :: Half -> Bool

Is this Half equal to 0?

fromHalf :: Half -> Float

Convert a Half to a Float while preserving NaN

toHalf :: Float -> Half

Convert a Float to a Half with proper rounding, while preserving NaN and dealing appropriately with infinity

Patterns

pattern POS_INF :: Half

Positive infinity

pattern NEG_INF :: Half

Negative infinity

pattern QNaN :: Half

Quiet NaN

pattern SNaN :: Half

Signalling NaN

pattern HALF_MIN :: Half

Smallest positive half

pattern HALF_NRM_MIN :: Half

Smallest positive normalized half

pattern HALF_MAX :: Half

Largest positive half

pattern HALF_EPSILON :: Half

Smallest positive e for which half (1.0 + e) != half (1.0)

pattern HALF_DIG :: (Eq a, Num a) => a

Number of base 10 digits that can be represented without change

pattern HALF_MIN_10_EXP :: (Eq a, Num a) => a

Minimum positive integer such that 10 raised to that power is a normalized half

pattern HALF_MAX_10_EXP :: (Eq a, Num a) => a

Maximum positive integer such that 10 raised to that power is a normalized half

Pure conversions

pure_floatToHalf :: Float -> Half

Naive pure-Haskell implementation of toHalf.

pure_halfToFloat :: Half -> Float

Naive pure-Haskell implementation of fromHalf.